Move MusicUtils class to utils package
