+++ app/src/androidTest/java/com/amoseui/music/MusicPlaybackStress.java
@@ -26,6 +26,8 @@ import android.support.test.runner.AndroidJUnit4;
 import android.util.Log;
 import android.view.KeyEvent;
 
+import com.amoseui.music.utils.MusicUtils;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
