+++ app/src/androidTest/java/com/amoseui/music/TestPlaylist.java
@@ -27,6 +27,8 @@ import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.view.KeyEvent;
 
+import com.amoseui.music.utils.MusicUtils;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
