+++ app/src/main/java/com/amoseui/music/ScanningProgress.java
@@ -27,6 +27,8 @@ import android.provider.MediaStore;
 import android.view.Window;
 import android.view.WindowManager;
 
+import com.amoseui.music.utils.MusicUtils;
+
 public class ScanningProgress extends Activity
 {
     private final static int CHECK = 0;
