+++ app/src/main/java/com/amoseui/music/WeekSelector.java
@@ -22,7 +22,8 @@ import android.os.Bundle;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
-import android.widget.Button;
+
+import com.amoseui.music.utils.MusicUtils;
 
 public class WeekSelector extends Activity
 {
